[WIP] New low-level proxying C API
Implement a new proxying API that is meant to be suitable both for proxying in
syscall implementations as well as for proxying arbitrary user work. Since the
system proxying queue is processed from so many locations, it is dangerous to
use it for arbitrary work that might take a lock or otherwise block and the work
sent to it has to be structured more like native signal handlers. To avoid this
limitation, the new API allows users to create their own proxying queues that
are processed only when the target thread returns to the JS event loop or when
the user explicitly requests processing.

In contrast to the existing proxying API, this new API:

 - Never drops tasks (except in the case of allocation failure). It grows the
   task queues as necessary instead.
 - Does not attempt to dynamically type or dispatch queued functions, but rather
   uses statically typed function pointers that take `void*` arguments. This
   simplifies both the API and implementation. Packing of varargs into
   dynamically typed function wrappers could easily be layered on top of this
   API.
 - Is less redundant. There is only one way to proxy work synchronously or
   asynchronously to any thread.
 - Is more general. It allows waiting for a task to be explicitly signaled as
   done in addition to waiting for the proxied function to return.
 - Uses arrays instead of linked lists for better data locality.
 - Has a more uniform naming convention.

A follow-up PR will reimplement the existing proxying API in terms of this new
API.