Add new `emscripten_thread_can_block` function and use it in `emscripten_futex_wait`. NFC
This new function returns true only on the main browser thread and not
on the main node runtime thread (unlike
`emscripten_is_main_browser_thread` which is an historical misnomer).

Using this new native function along with the intrisic
`__builtin_wasm_memory_atomic_wait32` we can now implement
`emscripten_futex_wait` in native code and only call out to JS for the
non-block/busy-looping case.

As a followup we could also move the non-block/busy-looping code
to the native side but doing it this way keep this change small
and focused.